refactor(rag): move public entrypoints and reorder API args
