+++ app/core_plugins/slack/rag.py
@@ -28,7 +28,7 @@ from app.lib.rag import (
 )
 from app.llm.providers import BaseChatProvider
 from app.models.drive import DriveFile
-from app.rag.utils import resolve_source_name  # internal helper for source-name resolution (residual; see #398)
+from app.rag.utils import resolve_source_name
 
 logger = get_logger(__name__)
 
@@ -130,10 +130,6 @@ async def answer_question(
     ResponseType. config.fallback_message is reserved ONLY for the case where the
     agent ran successfully but found no relevant chunks.
 
-    Note: unlike the previous fan-out implementation, a single per-file agent failure
-    now fails the entire answer with RETRIEVAL_ERROR_MESSAGE (strict mode via
-    agentic_retrieve_context).
-
     Args:
         session: Async SQLModel session.
         user_id: Bot owner (RLS scope).
@@ -165,7 +161,7 @@ async def answer_question(
         return NO_FILES_RESOLVED_MESSAGE, ResponseType.no_files_resolved
 
     try:
-        chunks = await agentic_retrieve_context(user_id, file_ids, question, agent_llm)
+        chunks = await agentic_retrieve_context(question, file_ids, user_id, agent_llm)
     except DriveFileNotFoundError as exc:
         logger.error("Slack agentic RAG: file not found user=%d files=%s: %s", user_id, file_ids, exc)
         return DRIVE_FILE_NOT_FOUND_MESSAGE, ResponseType.drive_file_not_found
