+++ tests/rag/test_retrieve_chunks.py
@@ -25,19 +25,19 @@ def _ctx(source: str, *contents: str) -> RAGContext:
 class TestRetrieveChunks:
     @pytest.mark.asyncio
     async def test_empty_file_ids_returns_empty(self) -> None:
-        result = await retrieve_chunks(1, [], "q", MagicMock())
+        result = await retrieve_chunks("q", [], 1, MagicMock())
         assert result == []
 
     @pytest.mark.asyncio
     async def test_maps_chunks_to_retrieved_chunks(self) -> None:
         with (
-            patch("app.lib.rag.validate_files_ready", new=AsyncMock()),
+            patch("app.rag.retrieval.validate_files_ready", new=AsyncMock()),
             patch(
-                "app.lib.rag.get_context_via_agent_with_isolated_session",
+                "app.rag.retrieval.get_context_via_agent_with_isolated_session",
                 new=AsyncMock(side_effect=[_ctx("a.pdf", "alpha"), _ctx("b.pdf", "beta")]),
             ),
         ):
-            result = await retrieve_chunks(1, [10, 11], "q", MagicMock())
+            result = await retrieve_chunks("q", [10, 11], 1, MagicMock())
         assert result == [
             RetrievedChunk(source_name="a.pdf", chapter="Ch", section=None, subsection=None, content="alpha"),
             RetrievedChunk(source_name="b.pdf", chapter="Ch", section=None, subsection=None, content="beta"),
@@ -48,13 +48,13 @@ class TestRetrieveChunks:
         get_ctx = AsyncMock()
         with (
             patch(
-                "app.lib.rag.validate_files_ready",
+                "app.rag.retrieval.validate_files_ready",
                 new=AsyncMock(side_effect=DriveFileNotFoundError("nope")),
             ),
-            patch("app.lib.rag.get_context_via_agent_with_isolated_session", new=get_ctx),
+            patch("app.rag.retrieval.get_context_via_agent_with_isolated_session", new=get_ctx),
         ):
             with pytest.raises(DriveFileNotFoundError):
-                await retrieve_chunks(1, [10], "q", MagicMock())
+                await retrieve_chunks("q", [10], 1, MagicMock())
         get_ctx.assert_not_awaited()  # nothing searched
 
     @pytest.mark.asyncio
@@ -62,23 +62,23 @@ class TestRetrieveChunks:
         get_ctx = AsyncMock()
         with (
             patch(
-                "app.lib.rag.validate_files_ready",
+                "app.rag.retrieval.validate_files_ready",
                 new=AsyncMock(side_effect=RAGNotReadyError(10, str(RagStatus.PROCESSING))),
             ),
-            patch("app.lib.rag.get_context_via_agent_with_isolated_session", new=get_ctx),
+            patch("app.rag.retrieval.get_context_via_agent_with_isolated_session", new=get_ctx),
         ):
             with pytest.raises(RAGNotReadyError):
-                await retrieve_chunks(1, [10], "q", MagicMock())
+                await retrieve_chunks("q", [10], 1, MagicMock())
         get_ctx.assert_not_awaited()
 
     @pytest.mark.asyncio
     async def test_retrieval_error_propagates(self) -> None:
         with (
-            patch("app.lib.rag.validate_files_ready", new=AsyncMock()),
+            patch("app.rag.retrieval.validate_files_ready", new=AsyncMock()),
             patch(
-                "app.lib.rag.get_context_via_agent_with_isolated_session",
+                "app.rag.retrieval.get_context_via_agent_with_isolated_session",
                 new=AsyncMock(side_effect=RAGRetrievalError("agent boom")),
             ),
         ):
             with pytest.raises(RAGRetrievalError):
-                await retrieve_chunks(1, [10], "q", MagicMock())
+                await retrieve_chunks("q", [10], 1, MagicMock())
